Reject commands from multiple transports
Apdu-dispatch does not have proper support for multiple concurrent
transport since there are no separate selection for each app.

The proper solution would be to have separate buffers and selection per transport,
but that would also require apps to actually deal with multiple interfaces, which they currently do not.

This patch makes it so that the first command sets the used transport, and that
 any command comming from another transport will be rejected.